+++ packages/migrations/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @swisspost/design-system-migrations
 
+## 9.0.0-next.18
+
+## 9.0.0-next.17
+
+## 9.0.0-next.16
+
 ## 9.0.0-next.15
 
 ## 9.0.0-next.14